requirements.txt
asgiref==3.2.10
attrs==20.2.0
beautifulsoup4==4.9.1
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
coverage==5.2.1
cssselect2==0.3.0
defusedxml==0.6.0
dj-stripe==2.3.0
Django==3.1.1
django-appconf==1.0.4
django-bootstrap4==2.2.0
django-cors-headers==3.5.0
django-countries==6.1.3
django-crispy-forms==1.9.2
django-formtools==2.2
django-inline-svg==0.1.1
django-jquery-js==3.1.1
django-js-asset==1.2.2
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-tinymce4-lite==1.8.0
django-weasyprint==1.0.1
djangorestframework==3.11.1
html5lib==1.1
idna==2.10
jsmin==2.2.2
jsonfield==3.1.0
Pillow==7.2.0
pycparser==2.20
PyPDF2==1.26.0
Pyphen==0.9.5
pytz==2020.1
reportlab==3.5.49
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
stripe==2.51.0
swapper==1.1.2.post1
tinycss2==1.0.2
urllib3==1.25.10
WeasyPrint==51
webencodings==0.5.1
